+++ source/CMakeLists.txt
@@ -11,7 +11,7 @@ file(GLOB cxx_headers                    ${CMAKE_CURRENT_LIST_DIR}/timemory/*.hp
 file(GLOB cxx_headers_variadic           ${CMAKE_CURRENT_LIST_DIR}/timemory/variadic/*.hpp)
 file(GLOB_RECURSE cxx_headers_backends   ${CMAKE_CURRENT_LIST_DIR}/timemory/backends/*.hpp)
 file(GLOB_RECURSE cxx_headers_components ${CMAKE_CURRENT_LIST_DIR}/timemory/components/*.hpp)
-file(GLOB cxx_headers_cereal        ${CMAKE_CURRENT_LIST_DIR}/timemory/cereal/*.hpp)
+file(GLOB_RECURSE cxx_headers_tpls       ${CMAKE_CURRENT_LIST_DIR}/timemory/tpls/*.hpp)
 file(GLOB cxx_headers_macros        ${CMAKE_CURRENT_LIST_DIR}/timemory/macros/*.hpp)
 file(GLOB cxx_headers_api           ${CMAKE_CURRENT_LIST_DIR}/timemory/api/*.hpp)
 file(GLOB cxx_headers_mpl           ${CMAKE_CURRENT_LIST_DIR}/timemory/mpl/*.hpp)
@@ -44,7 +44,7 @@ endforeach()
 set(timemory_glob_headers
     ${c_headers} ${cxx_headers} ${cxx_headers_variadic} ${cxx_headers_api}
     ${cxx_headers_backends} ${cxx_headers_components} ${cxx_headers_mpl}
-    ${cxx_headers_utility} ${cxx_headers_bits} ${cxx_headers_data} ${cxx_headers_cereal}
+    ${cxx_headers_utility} ${cxx_headers_bits} ${cxx_headers_data} ${cxx_headers_tpls}
     ${cxx_headers_category_bits} ${cxx_headers_general} ${cxx_headers_macros}
     ${compat_headers})
