+++ source/timemory/components/base/declaration.hpp
@@ -30,12 +30,12 @@
 
 #pragma once
 
-#include "timemory/cereal/cereal.hpp"
 #include "timemory/components/base/types.hpp"
 #include "timemory/components/properties.hpp"
 #include "timemory/mpl/types.hpp"
 #include "timemory/operations/types.hpp"
 #include "timemory/storage/types.hpp"
+#include "timemory/tpls/cereal.hpp"
 
 //======================================================================================//
