+++ source/timemory/components/base/definition.hpp
@@ -31,13 +31,13 @@
 
 #pragma once
 
-#include "timemory/cereal/cereal.hpp"
 #include "timemory/components/base/declaration.hpp"
 #include "timemory/components/base/templates.hpp"
 #include "timemory/components/base/types.hpp"
 #include "timemory/mpl/math.hpp"
 #include "timemory/mpl/types.hpp"
 #include "timemory/storage/types.hpp"
+#include "timemory/tpls/cereal.hpp"
 #include "timemory/units.hpp"
 
 //======================================================================================//
