+++ source/timemory/components/base/templates.hpp
@@ -31,7 +31,6 @@
 
 #pragma once
 
-#include "timemory/cereal/cereal.hpp"
 #include "timemory/components/base/declaration.hpp"
 #include "timemory/components/base/types.hpp"
 #include "timemory/mpl/math.hpp"
@@ -42,6 +41,7 @@
 #include "timemory/operations/types/base_printer.hpp"
 #include "timemory/operations/types/serialization.hpp"
 #include "timemory/storage/declaration.hpp"
+#include "timemory/tpls/cereal.hpp"
 #include "timemory/units.hpp"
 
 #include <cassert>
