+++ source/timemory/components/timing/ert_timer.hpp
@@ -24,9 +24,9 @@
 
 #pragma once
 
-#include "timemory/cereal/cereal.hpp"
 #include "timemory/components/timing/backends.hpp"
 #include "timemory/operations/types/base_printer.hpp"
+#include "timemory/tpls/cereal.hpp"
 #include "timemory/units.hpp"
 
 #include <cstdint>
