+++ source/timemory/data/statistics.hpp
@@ -33,11 +33,11 @@
 
 //----------------------------------------------------------------------------//
 
-#include "timemory/cereal/archives.hpp"
 #include "timemory/data/functional.hpp"
 #include "timemory/data/stream.hpp"
 #include "timemory/mpl/math.hpp"
 #include "timemory/mpl/stl.hpp"
+#include "timemory/tpls/cereal.hpp"
 #include "timemory/utility/macros.hpp"
 
 #include <cmath>
