+++ source/timemory/ert/counter.hpp
@@ -32,7 +32,6 @@
 
 #include "timemory/backends/device.hpp"
 #include "timemory/backends/dmp.hpp"
-#include "timemory/cereal/archives.hpp"
 #include "timemory/components/cuda/backends.hpp"
 #include "timemory/components/timing/ert_timer.hpp"
 #include "timemory/ert/aligned_allocator.hpp"
@@ -42,6 +41,7 @@
 #include "timemory/ert/types.hpp"
 #include "timemory/mpl/apply.hpp"
 #include "timemory/settings/declaration.hpp"
+#include "timemory/tpls/cereal/archives.hpp"
 #include "timemory/utility/macros.hpp"
 
 #include <array>
