+++ source/timemory/operations/types/finalize/get.hpp
@@ -30,7 +30,6 @@
 #pragma once
 
 #include "timemory/backends/dmp.hpp"
-#include "timemory/cereal/cereal.hpp"
 #include "timemory/manager/declaration.hpp"
 #include "timemory/operations/declaration.hpp"
 #include "timemory/operations/macros.hpp"
@@ -38,6 +37,7 @@
 #include "timemory/storage/basic_tree.hpp"
 #include "timemory/storage/node.hpp"
 #include "timemory/storage/types.hpp"
+#include "timemory/tpls/cereal.hpp"
 
 #include <string>
 #include <vector>
