+++ source/timemory/operations/types/serialization.hpp
@@ -29,10 +29,10 @@
 
 #pragma once
 
-#include "timemory/cereal/archives.hpp"
 #include "timemory/operations/declaration.hpp"
 #include "timemory/operations/macros.hpp"
 #include "timemory/operations/types.hpp"
+#include "timemory/tpls/cereal/archives.hpp"
 
 namespace tim
 {
