+++ source/timemory/settings/settings.cpp
@@ -24,10 +24,10 @@
 
 #include "timemory/settings/settings.hpp"
 #include "timemory/backends/dmp.hpp"
-#include "timemory/cereal/archives.hpp"
 #include "timemory/mpl/policy.hpp"
 #include "timemory/settings/macros.hpp"
 #include "timemory/settings/types.hpp"
+#include "timemory/tpls/cereal/archives.hpp"
 #include "timemory/utility/argparse.hpp"
 #include "timemory/utility/bits/signals.hpp"
 #include "timemory/utility/declaration.hpp"
