+++ source/timemory/settings/tsettings.hpp
@@ -24,9 +24,9 @@
 
 #pragma once
 
-#include "timemory/cereal/cereal.hpp"
 #include "timemory/settings/types.hpp"
 #include "timemory/settings/vsettings.hpp"
+#include "timemory/tpls/cereal.hpp"
 #include "timemory/utility/argparse.hpp"
 #include "timemory/utility/types.hpp"
