+++ source/timemory/storage/basic_tree.hpp
@@ -24,8 +24,8 @@
 
 #pragma once
 
-#include "timemory/cereal/cereal.hpp"
 #include "timemory/hash/types.hpp"
+#include "timemory/tpls/cereal.hpp"
 #include "timemory/utility/types.hpp"
 
 #include <iterator>
