+++ source/timemory/storage/declaration.hpp
@@ -32,7 +32,6 @@
 #include "timemory/backends/dmp.hpp"
 #include "timemory/backends/gperf.hpp"
 #include "timemory/backends/threading.hpp"
-#include "timemory/cereal/cereal.hpp"
 #include "timemory/hash/declaration.hpp"
 #include "timemory/manager/declaration.hpp"
 #include "timemory/mpl/policy.hpp"
@@ -45,6 +44,7 @@
 #include "timemory/storage/macros.hpp"
 #include "timemory/storage/node.hpp"
 #include "timemory/storage/types.hpp"
+#include "timemory/tpls/cereal.hpp"
 #include "timemory/utility/macros.hpp"
 #include "timemory/utility/singleton.hpp"
 #include "timemory/utility/types.hpp"
