+++ source/timemory/storage/node.hpp
@@ -26,10 +26,10 @@
 
 #include "timemory/backends/process.hpp"
 #include "timemory/backends/threading.hpp"
-#include "timemory/cereal/archives.hpp"
 #include "timemory/hash.hpp"
 #include "timemory/mpl/type_traits.hpp"
 #include "timemory/mpl/types.hpp"
+#include "timemory/tpls/cereal/archives.hpp"
 
 #include <cstdint>
 #include <set>
