+++ source/timemory/tpls/cereal/archives.hpp
@@ -24,7 +24,8 @@
 
 #pragma once
 
-#include "timemory/cereal/cereal.hpp"
+#include "timemory/tpls/cereal/cereal.hpp"
+#include "timemory/tpls/cereal/types.hpp"
 
 // cereal will cause some -Wclass-memaccess warnings that are quite annoying
 #if defined(__GNUC__) && (__GNUC__ > 7)
@@ -32,16 +33,11 @@
 #    pragma GCC diagnostic ignored "-Wclass-memaccess"
 #endif
 
-// types
-#include "cereal/types/array.hpp"
-#include "cereal/types/common.hpp"
-#include "cereal/types/map.hpp"
-#include "cereal/types/memory.hpp"
-#include "cereal/types/set.hpp"
-#include "cereal/types/string.hpp"
-#include "cereal/types/tuple.hpp"
-#include "cereal/types/utility.hpp"
-#include "cereal/types/vector.hpp"
+// archives
+#include "cereal/archives/json.hpp"
+#if defined(TIMEMORY_USE_XML_ARCHIVE)
+#    include "cereal/archives/xml.hpp"
+#endif
 
 #if defined(__GNUC__) && (__GNUC__ > 7)
 #    pragma GCC diagnostic pop
