+++ source/timemory/variadic/component_list.hpp
@@ -35,13 +35,13 @@
 #pragma once
 
 #include "timemory/backends/dmp.hpp"
-#include "timemory/cereal/cereal.hpp"
 #include "timemory/general/source_location.hpp"
 #include "timemory/mpl/apply.hpp"
 #include "timemory/mpl/filters.hpp"
 #include "timemory/operations/types.hpp"
 #include "timemory/settings/declaration.hpp"
 #include "timemory/storage/types.hpp"
+#include "timemory/tpls/cereal.hpp"
 #include "timemory/utility/macros.hpp"
 #include "timemory/variadic/base_bundle.hpp"
 #include "timemory/variadic/functional.hpp"
